qtbase: Package the 3rd party libs if built
Third-party libs e.g. libqtfreetype when built is not packaged and hence
goes missing in sysroot, then rest of qt components like qtwayland gets
its configuration from qtbase to enable it but then does not find the
internal version of libqtfreetype.a which is what qtbase directs it to
use and as a result it starts throwing linking errors e.g.

x86_64-angstrom-linux-ld.gold: error: cannot find -lqtfreetype

This patch installs the needed libs into dev packages

Remove a trailing whitespace as well while at it.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Tested-by: Denys Dmytriyenko <denys@ti.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>